Fixed pathes for filesystem

+++ bin/src/filesystem/file/file.service.ts
@@ -5,7 +5,12 @@ import { File } from '../models/file.model';
 
 export const remove = async (req: Request, res: Response) => {
   try {
-    const filePath = extractPath(req.path, fsCommand.rm);
+    const userDir = req.user.sub;
+    if (!userDir) {
+      return res.status(404).json({ error: 'Forbidden' });
+    }
+
+    const filePath = extractPath(req.path, userDir, fsCommand.rm);
     await File.remove(filePath);
     return res.status(200).json({ message: 'File removed.' });
   } catch (error) {
